% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataculture-package.R
\docType{package}
\name{dataculture-package}
\alias{dataculture}
\alias{dataculture-package}
\title{dataculture: Supporting Materials for Data and Culture Course}
\description{
Provides example cultural datasets as used in the R labs in Data & Culture 359:207 at Rutgers-New Brunswick in Fall 2022.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agoldst.github.io/dataculture/}
}

}
\author{
\strong{Maintainer}: Andrew Goldstone \email{andrew.goldstone@rutgers.edu}

}
\keyword{internal}
