% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genre_features}
\alias{genre_features}
\title{Document-term matrix for fiction genres from Underwood, Distant Horizons, chap. 2}
\format{
\subsection{\code{genre_features}}{

A sparse dgCMatrix with 1047 rows corresponding to documents and 4889 columns corresponding to features. The row names are document IDs corresponding to \code{genre_meta$docid} and the column names given the features as used by Underwood.
}
}
\source{
\url{https://www.ideals.illinois.edu/items/105528} for the data; files in \url{https://github.com/tedunderwood/horizon/blob/master/chapter2/modeloutput} for the choice of features.
}
\usage{
genre_features
}
\description{
Data (extracted word-frequency and other features for fiction volumes) used in modeling fiction genre by Underwood in \emph{Distant Horizons}, chap. 2. The features are only those used in the regularized logistic regression models of science fiction, detective fiction, and Gothic supplied by Underwood in his reproduction repository, and do not include all features found in the accompanying source data files.
}
\keyword{datasets}
