% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genre_meta}
\alias{genre_meta}
\title{Volume Metadata from Underwood, Distant Horizons, Chap. 2}
\format{
\subsection{\code{genre_meta}}{

A data frame with 1047 rows and 15 columns, notably
\describe{
\item{docid}{unique identifier corresponding to extracted features filename}
\item{firstpub}{year of first publication of title}
\item{tags}{string with genre tags, separated by \code{" | "}}
}
}
}
\source{
\url{https://github.com/tedunderwood/horizon/blob/master/chapter2/metadata/concatenatedmeta.csv}; for the description of genre tags, \url{https://github.com/tedunderwood/horizon/blob/master/chapter2/metadata/genrecategorieschapter2.docx}; and see Ted Underwood, \emph{Distant Horizons: Digital Evidence and Literary Change} (Chicago: University of Chicago Press, 2019), chap. 2.
}
\usage{
genre_meta
}
\description{
Metadata from Ted Underwood's reproduction repository for \emph{Distant Horizons}, chap. 2: publication information, record locators, and genre tags derived from bibliographic sources for a selection of science-fiction, detective, Gothic, and "random" fiction.
}
\details{
Underwood's information about the \code{tags} column:\tabular{llll}{
   tag \tab texts \tab dates \tab description or source \cr
   det100 \tab 89 \tab 1829 - 1941 \tab \emph{The First Hundred Years of Detective Fiction, 1841-1941.} 1973. Lilly Library, Bloomington, IN. \url{http://www.indiana.edu/~liblilly/etexts/detective/} \cr
   chimyst \tab 146 \tab 1923-1989 \tab Works categorized by librarians as “detective” or “mystery fiction,” collected at the Chicago Text Lab. \cr
   locdetmyst \tab 45 \tab 1832-1922 \tab Works categorized by librarians as “detective and mystery fiction,” collected in HathiTrust. \cr
   locdetective \tab 16 \tab 1865-1912 \tab Works categorized by librarians with the subject heading “Detectives.” Often casebook fiction. \cr
   crime \tab 2 \tab 1972-1974 \tab Works categorized by librarians as “crime fiction” but not “detective fiction.” \cr
   cozy \tab 10 \tab 1920-1952 \tab Works by authors mentioned as writing country-house mysteries in \emph{The Mystery Readers' Advisory: The Librarian's Clues to Murder and Mayhem}, by John Charles, Joanna Morrison, and Candace Clark (Chicago: ALA, 2002). \cr
   hardboiled \tab 10 \tab 1929-1970 \tab Appendix to Geoffrey O’Brien, \emph{Hardboiled America} (New York: Van Nostrand Reinhold, 1981). \cr
   newgate \tab 7 \tab 1828-1840 \tab Keith Hollingsworth, \emph{The Newgate Novel,} Detroit 1963. \cr
   sensation \tab 14 \tab 1860-1880 \tab “The Sensation Novel,” Winifred Hughes, in \emph{A Companion to the Victorian Novel,} ed. Patrick Brantlinger and William B. Thesing (Blackwell, 2002). \cr
   lockandkey \tab 10 \tab 1800-1903 \tab Works anthologized in \emph{The Lock and Key Library: Classic Mystery and Detective Stories,} edited by Julian Hawthorne (New York, 1909). Includes writers like Dostoevsky who probably were not seen as writers of mysteries even in 1909; not included in the article’s model of “detective fiction.” \cr
   pbgothic \tab 96 \tab 1764-1988 \tab “Chronology” in David Punter and Glennis Byron, \emph{The Gothic} (Malden: Blackwell, 2004). \cr
   stangothic \tab 21 \tab 1791-1834 \tab A small subset of works tagged as “Gothic” in Stanford Literary Lab metadata. \cr
   lochorror \tab 4 \tab 1818 \tab Works tagged as “horror” by librarians, collected in HathiTrust. \cr
   chihorror \tab 23 \tab 1933-1989 \tab Works tagged as “horror” by librarians, collected in the Chicago Text Lab. \cr
   locghost \tab 28 \tab 1826-1922 \tab Works tagged as “ghost stories” by librarians. \cr
   locscifi \tab 21 \tab 1836-1909 \tab Works tagged as “science fiction” by librarians and collected in HathiTrust. \cr
   chiscifi \tab 144 \tab 1901-1989 \tab Works tagged as “science fiction” by librarians and collected at the Chicago Text Lab. \cr
   femscifi \tab 9 \tab 1818-1922 \tab Ockerbloom, Mary Mark. 2015. “Pre-1950 Utopias and Science Fiction by Women.” \cr
   anatscifi \tab 36 \tab 1771-1922 \tab Stableford, Brian. 2004. “The Emergence of Science Fiction” and “Science Fiction Between the Wars.” In \emph{Anatomy of Wonder,} edited by Neil Barron, 5th edition, 3-44. \cr
   chiutopia \tab 13 \tab 1920-1976 \tab Works tagged as “Utopias” by librarians, collected at the Chicago Text Lab, not folded into “science fiction” in this article. \cr
   chifantasy \tab 53 \tab 1901-1989 \tab Works tagged as “fantastic” or “fantasy fiction” by librarians, not folded into “science fiction” for the purposes of this article. \cr
   juvenile \tab 23 \tab 1904-1922 \tab Works for a juvenile audience; collected but not used in this article. \cr
   drop \tab 33 \tab 1838-1922 \tab Works that I decided not to use, left in the metadata for transparency. The most common reason is that they are juvenile. \cr
   random \tab 169 \tab 1769-1922 \tab Works randomly selected from HathiTrust Digital Library, using fiction metadata developed in the NEH-funded project “Understanding Genre in a Collectiono f a Million Volumes.” “Random selection” here means that the volumes were selected randomly but then approved or rejected by the author, to avoid stray volumes of nonfiction, classical poetry, juvenile works, etc. \cr
   chirandom \tab 202 \tab 1920-1989 \tab Works randomly selected from the Chicago Text Lab. Selection here was more genuinely random. Note that both “random” tags can coexist with other genre tags. A randomly-selected volume could also be “chimyst,” for instance; in that case it will be excluded from the negative (contrast) set only if “chimyst” is in the positive set. \cr
   teamred \tab 484 \tab 1760-1989 \tab Randomly selected authors for a sanity check. \cr
   teamblack \tab 500 \tab 1764-1989 \tab Randomly selected authors for a sanity check. \cr
   stew \tab 224 \tab 1764-1989 \tab A random selection of volumes balanced between Gothic, science fiction, and crime/detective traditions, in order to create a ghastly genre stew. \cr
}
}
\keyword{datasets}
