% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gssmusic}
\alias{gssmusic}
\title{General Social Survey 1993 Culture Module Excerpt}
\format{
\subsection{\code{gssmusic}}{

A data frame with 1606 rows and 33 columns.
\describe{
\item{music_likes}{number of music genres respondent likes or likes very much}
\item{music_dislikes}{number of music genres respondent dislikes or dislikes very much}
}
}
}
\source{
General Social Survey, \url{https://gss.norc.org/}, via Kieran Healy's gssr package \url{https://kjhealy.github.io/gssr/}.
}
\usage{
gssmusic
}
\description{
Selected variables from the 1993 General Social Survey, including the questions on musical taste from the Culture Module. Data has been selected from the Cumulative Data File made available in the \code{gssr} package, categorical variables have been converted to \code{factor} type for convenience, and two new variables, \code{music_likes} and \code{music_dislikes}, have been created.
}
\keyword{datasets}
