% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamlet}
\alias{hamlet}
\title{Speaker interactions in Hamlet}
\format{
\subsection{\code{hamlet}}{

A data frame with 140 rows and 5 columns
\describe{
\item{s1}{first speaker in interaction, alphabetically}
\item{s2}{second speaker in interaction, alphabetically}
\item{act}{act number of the play in which the pair appears}
\item{scene}{scene number of the play in which the pair appears}
\item{turns}{number of times either speaks after the other in that act and scene}
}
}
}
\source{
\url{https://shakespeare.folger.edu/shakespeares-works/hamlet/download/}. The New Folger Shakespeare is available under a \href{https://creativecommons.org/licenses/by-nc/3.0/deed.en_US}{CC-BY-NC 3.0} license. The algorithm for extracting interactions is based on Folgert Karsdorp et al., \emph{Humanities Data Analysis: Case Studies with Python} (Princeton University Press, 2021), chap. 2.
}
\usage{
hamlet
}
\description{
Speaker interactions in Shakespeare's \emph{Hamlet}. Two characters are recorded as interacting if one speaks directly after another within a scene of the play. X speaking after Y is counted the same as Y speaking after X. Data has been extracted from the XML version of the New Folger Library Edition of \emph{Hamlet} (a conflated text of the play) following a method demonstrated in Karsdorp et al.'s \emph{Humanities Data Analysis}.
}
\details{
Speaker identifications have been extracted from \code{who} attributes of \code{sp} speech-heading tags, with leading \verb{#} and trailing \verb{_Ham} removed.
}
\keyword{datasets}
