% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ingreds}
\alias{ingreds}
\title{Historic American Cookbook Recipes: Ingredients Data}
\format{
\subsection{\code{ingreds}}{

A data frame with 382487 rows and 7 columns:
\describe{
\item{recipe_id}{sequential recipe number}
\item{book_id}{book identifier (filename in Feeding America data)}
\item{date}{year of publication}
\item{ethnicgroup}{ethnic category (may be \code{NA})}
\item{recipe_class}{type of recipe}
\item{region}{geographic region of origin}
\item{ingredients}{name of ingredient}
}
}
}
\source{
\url{https://doi.org/10.5281/zenodo.891264}
}
\usage{
ingreds
}
\description{
Unnested version of \code{recipes}.
}
\details{
For exploratory convenience, a duplicate version of \code{recipes} with the
\code{ingredients} list-column unnested and each recipe ssigned a distinct ID.
}
\keyword{datasets}
