% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recent_ssa_names}
\alias{recent_ssa_names}
\title{US baby names, 2016–2020}
\format{
\subsection{\code{recipes}}{

A data frame with 161,413 rows and 5 columns:
\describe{
\item{year}{birth year}
\item{name}{first name}
\item{sex}{\code{M} or \code{F}}
\item{n}{count}
\item{prop}{frequency per live births of that sex}
}
}
}
\source{
\url{https://www.ssa.gov/oact/babynames/names.zip} for the names, \url{http://wonder.cdc.gov/natality-expanded-current.html} for live births data.
}
\usage{
recent_ssa_names
}
\description{
Names by birth year and sex, from the U.S. Social Security Administration.
}
\details{
A supplement to the \code{babynames} package with a few more years of data. Frequencies are not strictly comparable as the denominator is US live births rather than Social Security applicants as in \code{babynames}.
}
\keyword{datasets}
